api/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
psycopg2-binary==2.9.10
databases[postgresql]==0.9.0
sqlalchemy==2.0.43
pika==1.3.2